Improve filename generation: restore random suffix and simplify format
- Restore random suffix functionality for unique filenames (configurable)
- Remove date suffix from filenames for cleaner names
- Remove document type from filename (redundant with metadata)
- Fix filename duplication issue where type and title were both included
- New format: title.randomSuffix.typ (e.g., meeting.A7bX9K.typ)
- Add configuration options: use_random_suffix, random_suffix_length
- Maintain backwards compatibility with doc_type parameter
